+++ src/synth.h
@@ -1,32 +1,32 @@
 #ifndef __SYNTH_H__
 #define __SYNTH_H__
 
-float synth_callback(double time);
+double synth_callback(double time);
 
-float sine_oscillator(double time, float frequency, float amplitude);
+double sine_oscillator(double time, double frequency, double amplitude);
 
-float square_oscillator(
+double square_oscillator(
     double time,
-    float frequency,
-    float amplitude,
+    double frequency,
+    double amplitude,
     int num_harmonics
 );
 
-float sawtooth_oscillator(
+double sawtooth_oscillator(
     double time,
-    float frequency,
-    float amplitude,
+    double frequency,
+    double amplitude,
     int num_harmonics
 );
 
 
-float triangle_oscillator(
+double triangle_oscillator(
     double time,
-    float frequency,
-    float amplitude,
+    double frequency,
+    double amplitude,
     int num_harmonics
 );
 
-float white_noise(float amplitude);
+double white_noise(double amplitude);
 
 #endif
